Move `RenderPass*AttachmentDescriptor` types into core
