+++ wgpu-types/src/lib.rs
@@ -1077,74 +1077,6 @@ pub enum SwapChainStatus {
     OutOfMemory,
 }
 
-/// Operation to perform to the output attachment at the start of a renderpass.
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum LoadOp {
-    /// Clear the output attachment with the clear color. Clearing is faster than loading.
-    Clear = 0,
-    /// Do not clear output attachment.
-    Load = 1,
-}
-
-/// Operation to perform to the output attachment at the end of a renderpass.
-#[repr(C)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum StoreOp {
-    /// Clear the render target. If you don't care about the contents of the target, this can be faster.
-    Clear = 0,
-    /// Store the result of the renderpass.
-    Store = 1,
-}
-
-/// Describes an individual channel within a render pass, such as color, depth, or stencil.
-#[repr(C)]
-#[derive(Clone, Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct PassChannel<V> {
-    /// Operation to perform to the output attachment at the start of a renderpass. This must be clear if it
-    /// is the first renderpass rendering to a swap chain image.
-    pub load_op: LoadOp,
-    /// Operation to perform to the output attachment at the end of a renderpass.
-    pub store_op: StoreOp,
-    /// If load_op is [`LoadOp::Clear`], the attachement will be cleared to this color.
-    pub clear_value: V,
-    /// If true, the relevant channel is not changed by a renderpass, and the corresponding attachment
-    /// can be used inside the pass by other read-only usages.
-    pub read_only: bool,
-}
-
-/// Describes a color attachment to a [`RenderPass`].
-#[repr(C)]
-#[derive(Clone, Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct RenderPassColorAttachmentDescriptorBase<T> {
-    /// Texture attachment to render to. Must contain [`TextureUsage::OUTPUT_ATTACHMENT`].
-    pub attachment: T,
-    /// MSAA resolve target. Must contain [`TextureUsage::OUTPUT_ATTACHMENT`]. Must be `None` if
-    /// attachment has 1 sample (does not have MSAA). This is not mandatory for rendering with multisampling,
-    /// you can choose to resolve later or manually.
-    pub resolve_target: Option<T>,
-    /// Color channel.
-    pub channel: PassChannel<Color>,
-}
-
-/// Describes a depth/stencil attachment to a [`RenderPass`].
-#[repr(C)]
-#[derive(Clone, Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct RenderPassDepthStencilAttachmentDescriptorBase<T> {
-    /// Texture attachment to render to. Must contain [`TextureUsage::OUTPUT_ATTACHMENT`] and be a valid
-    /// texture type for a depth/stencil attachment.
-    pub attachment: T,
-    /// Depth channel.
-    pub depth: PassChannel<f32>,
-    /// Stencil channel.
-    pub stencil: PassChannel<u32>,
-}
-
 /// Describes the attachments of a render pass.
 #[derive(Clone, Debug, Default, PartialEq)]
 pub struct RenderPassDescriptor<'a, C, D> {
